requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
redis[asyncio]==5.0.8
python-dotenv==1.0.1
pydantic==2.8.2
psycopg2-binary==2.9.9
pandas==2.1.3
openpyxl==3.1.5
langchain==0.3.26
langchain-core==0.3.68
langchain-community==0.3.27
langchain-google-genai==2.1.8
langgraph==0.6.4
google-auth==2.40.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.105.0
